fix: change resetPassword parameter type to String
This function is called before the user is logged, thus making use of the User Model will require all the fields to be instantiated, where as only the email is required, which is a string.

Also enforced named paramaters for primitive types to avoid passing in wrong values or types.